feat: update login modal and fix bind modal
