--- src/components/BindDidModal/Home/index.module.scss
@@ -1,47 +1,0 @@
-@import "../../../style/index.scss";
-
-.chooseWalletBox {
-  .contentBox {
-    padding: 16px 0;
-    @include themeify {
-      color: themed(Gray_900);
-
-      .top {
-        display: flex;
-        justify-content: left;
-        align-items: center;
-        font-family: 'Inter', sans-serif;
-        font-style: normal;
-        line-height: 20px;
-        text-align: center;
-        height: 20px;
-        font-weight: themed(Regular_font-weight);
-        font-size: themed(Body_font-size);
-
-        .icon {
-          margin-right: 4px;
-          width: 16px;
-          height: 16px;
-        }
-      }
-
-      .btnsBox {
-        .btn {
-          width: 100%;
-          margin-top: 16px;
-          height: 40px;
-
-          span {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-          }
-
-          .icon {
-            margin-right: 8px;
-          }
-        }
-      }
-    }
-  }
-}
