--- src/components/BindDidModal/RenderWallets/index.ts
@@ -1,1 +1,0 @@
-export * from './RenderWallets';
