+++ src/components/LoginModal/Home/Home.tsx
@@ -1,68 +1,49 @@
-import React from 'react';
+import React, {ReactNode} from 'react';
 import { DesktopIcon, MobileIcon, Web3MqIcon } from '../../../icons';
 import { Button } from '../../Button';
-import { RenderWallets } from '../RenderWallets';
-import { StepStringEnum, useLoginContext } from '../../../context';
 import ss from './index.module.scss';
 import cx from 'classnames';
-import { DappConnect } from '@web3mq/dapp-connect';
 
 type IProps = {
-  WalletConnectBtnNode: React.ReactNode;
+  styles: Record<string, any> | null;
+  WalletConnectBtnNode: ReactNode;
+  handleWeb3MQClick: () => void;
+  RenderWallets: ReactNode
 };
 export const Home: React.FC<IProps> = (props) => {
-  const {WalletConnectBtnNode } = props;
-  const { step, styles, setStep, env, setDappConnectClient } = useLoginContext();
-
-  const handleWeb3mqClick = () => {
-    new Promise((resolve) => {
-      setDappConnectClient(
-        new DappConnect({ dAppID: 'SwapChat:im', keepAlive: false, env }, () => {}),
-      );
-      resolve('success');
-    }).then(() => {
-      setStep(StepStringEnum.QR_CODE);
-    });
-  };
+  const { WalletConnectBtnNode, handleWeb3MQClick, styles, RenderWallets } = props;
 
   return (
     <div className={cx(ss.container)} style={styles?.homeContainer}>
-      {step === StepStringEnum.HOME && (
-        <div className={ss.chooseWalletBox}>
-          <div className={ss.contentBox} style={styles?.contentBox}>
-            <div className={ss.top}>
-              <div className={ss.icon}>
-                <DesktopIcon />
-              </div>
-              <div>Desktop</div>
+      <div className={ss.chooseWalletBox}>
+        <div className={ss.contentBox} style={styles?.contentBox}>
+          <div className={ss.top}>
+            <div className={ss.icon}>
+              <DesktopIcon />
             </div>
-            <RenderWallets showCount={3} />
+            <div>Desktop</div>
           </div>
-          <div className={ss.contentBox} style={styles?.contentBox}>
-            <div className={ss.top}>
+          { RenderWallets }
+          {/*<RenderWallets showCount={3} />*/}
+        </div>
+        <div className={ss.contentBox} style={styles?.contentBox}>
+          <div className={ss.top}>
+            <div className={ss.icon}>
+              <MobileIcon />
+            </div>
+            <div className={ss.title}>Mobile</div>
+          </div>
+          <div className={ss.btnsBox}>
+            <Button className={ss.btn} style={styles?.homeButton} onClick={handleWeb3MQClick}>
               <div className={ss.icon}>
-                <MobileIcon />
+                <Web3MqIcon />
               </div>
-              <div className={ss.title}>Mobile</div>
-            </div>
-            <div className={ss.btnsBox}>
-              <Button className={ss.btn} style={styles?.homeButton} onClick={handleWeb3mqClick}>
-                <div className={ss.icon}>
-                  <Web3MqIcon />
-                </div>
-                Web3MQ
-              </Button>
-              { WalletConnectBtnNode }
-              {/*<Button className={ss.btn} style={styles?.homeButton}>*/}
-              {/*  <div className={ss.icon}>*/}
-              {/*    <WalletConnectIcon />*/}
-              {/*  </div>*/}
-              {/*  WalletConnect*/}
-              {/*</Button>*/}
-            </div>
+              Web3MQ
+            </Button>
+            {WalletConnectBtnNode}
           </div>
         </div>
-      )}
+      </div>
     </div>
   );
 };
