+++ src/components/LoginModal/hooks/useLogin.tsx
@@ -1,7 +1,12 @@
-import { useRef, useState } from 'react';
+import React, { useRef, useState } from 'react';
 import type { WalletType } from '@web3mq/client';
-import { AppTypeEnum, SignAuditTypeEnum } from '../../../context';
+import { AppTypeEnum } from '../../../context';
 import type { DappConnect } from '@web3mq/dapp-connect';
+import SignClient from '@walletconnect/sign-client';
+import type { SessionTypes } from '@walletconnect/types';
+import * as encoding from '@walletconnect/encoding';
+import { Web3Modal } from '@web3modal/standalone';
+import { SignAuditTypeEnum } from '../../../types/enum';
 
 export type LoginEventType = 'login' | 'register' | 'error';
 export type LoginEventDataType = {
@@ -39,22 +44,252 @@ export type RegisterResType = {
   address: string;
 };
 
-// keys: for login with main keys
-// account: For custom get users
-const useLogin = (
-  handleLoginEvent: (eventData: LoginEventDataType) => void,
-  client: any,
-  dappConnectClient?: DappConnect,
-  keys?: MainKeysType,
-  account?: UserAccountType,
-  appType?: AppTypeEnum,
-) => {
+type IProps = {
+  client: any;
+  dappConnectClient?: DappConnect;
+  walletConnectClient: React.MutableRefObject<SignClient | undefined>;
+  handleLoginEvent: (eventData: LoginEventDataType) => void;
+  keys?: MainKeysType;
+  account?: UserAccountType;
+  appType?: AppTypeEnum;
+  propWcSession?: SessionTypes.Struct;
+};
+
+const projectId = '1c5ee52c12a9145d5b184e06120462cc';
+const relay = 'wss://relay.walletconnect.com';
+
+const web3Modal = new Web3Modal({
+  projectId,
+  walletConnectVersion: 2,
+  // `standaloneChains` can also be specified when calling `web3Modal.openModal(...)` later on.
+  standaloneChains: ['eip155:1'],
+});
+const useLogin = (props: IProps) => {
+  const {
+    client,
+    dappConnectClient,
+    walletConnectClient,
+    handleLoginEvent,
+    keys,
+    account,
+    appType,
+    propWcSession,
+  } = props;
+  const [wcSession, setWcSession] = useState<SessionTypes.Struct | undefined>(propWcSession);
   const [userAccount, setUserAccount] = useState<UserAccountType | undefined>(account);
   const confirmPassword = useRef<string>('');
   const walletAddress = useRef<string>('');
   const [registerSignRes, setRegisterSignRes] = useState('');
   const [mainKeys, setMainKeys] = useState<MainKeysType | undefined>(keys);
   const signType = useRef<SignAuditTypeEnum>();
+
+  const _subscribeToEvents = async (_client: SignClient) => {
+    if (typeof _client === 'undefined') {
+      throw new Error('WalletConnect is not initialized');
+    }
+
+    _client.on('session_ping', (args) => {
+      console.log('EVENT', 'session_ping', args);
+    });
+
+    _client.on('session_event', (args) => {
+      console.log('EVENT', 'session_event', args);
+    });
+
+    _client.on('session_update', ({ topic, params }) => {
+      console.log('EVENT', 'session_update', { topic, params });
+      const { namespaces } = params;
+      const _session = _client.session.get(topic);
+      const updatedSession = { ..._session, namespaces };
+      onSessionConnected(updatedSession);
+    });
+
+    _client.on('session_delete', () => {
+      console.log('EVENT', 'session_delete');
+    });
+  };
+
+  const _checkPersistedState = async (_client: SignClient) => {
+    if (typeof _client === 'undefined') {
+      throw new Error('WalletConnect is not initialized');
+    }
+    // populates existing pairings to state
+    const curPairing = _client.pairing.getAll({ active: true });
+    // setPairings(curPairing);
+    console.log('RESTORED PAIRINGS: ', _client.pairing.getAll({ active: true }));
+
+    if (typeof wcSession !== 'undefined') return;
+    // populates (the last) existing session to state
+    if (_client.session.length) {
+      const lastKeyIndex = _client.session.keys.length - 1;
+      const _session = _client.session.get(_client.session.keys[lastKeyIndex]);
+      console.log('RESTORED SESSION:', _session);
+      onSessionConnected(_session);
+    }
+  };
+
+  const create = async () => {
+    if (walletConnectClient.current) return;
+    walletConnectClient.current = await SignClient.init({
+      projectId,
+      // optional parameters
+      relayUrl: relay,
+      metadata: {
+        name: 'react component',
+        description: 'react component',
+        url: window.location.host,
+        icons: [],
+      },
+    });
+    await _subscribeToEvents(walletConnectClient.current);
+    // await _checkPersistedState(walletConnectClient.current);
+  };
+
+  const connect = async () => {
+    if (typeof walletConnectClient.current === 'undefined') {
+      throw new Error('WalletConnect is not initialized');
+    }
+    const { uri, approval } = await walletConnectClient.current.connect({
+      requiredNamespaces: {
+        eip155: {
+          methods: ['personal_sign'],
+          chains: ['eip155:1'],
+          events: ['chainChanged', 'accountsChanged'],
+        },
+      },
+    });
+
+    // Open QRCode modal if a URI was returned (i.e. we're not connecting an existing pairing).
+    if (uri) {
+      // Create a flat array of all requested chains across namespaces.
+      web3Modal.openModal({ uri });
+    }
+    const session = await approval();
+    return session;
+  };
+
+  const closeModal = () => {
+    web3Modal.closeModal();
+  };
+
+  const onSessionConnected = (_session: SessionTypes.Struct) => {
+    setWcSession(_session);
+  };
+
+  const sendSignByWalletConnect = async (signContent: string, address: string) => {
+    // encode message (hex)
+    const hexMsg = encoding.utf8ToHex(signContent, true);
+    // personal_sign params
+    const params = [hexMsg, address];
+    // send message
+    const signature = await walletConnectClient.current?.request({
+      topic: wcSession?.topic as string,
+      chainId: 'eip155:1',
+      request: {
+        method: 'personal_sign',
+        params,
+      },
+    });
+    return {
+      signature,
+    };
+  };
+
+  const getMainKeypairByWalletConnect = async (options: {
+    password: string;
+    did_value: string;
+    did_type: string;
+  }) => {
+    const { password, did_value, did_type } = options;
+    const { signContent } = await client.register.getMainKeypairSignContent({
+      password: password,
+      did_value,
+      did_type,
+    });
+    const { signature } = await sendSignByWalletConnect(signContent, did_value.toLowerCase());
+    const { publicKey, secretKey } = await client.register.getMainKeypairBySignature(
+      signature,
+      confirmPassword.current,
+    );
+    return { publicKey, secretKey };
+  };
+
+  const getRegisterSignContentByWalletConnect = async (options: {
+    userid: string;
+    mainPublicKey: string;
+    didType: string;
+    didValue: string;
+  }) => {
+    const { userid, mainPublicKey, didValue, didType } = options;
+    const { signContent } = await client.register.getRegisterSignContent({
+      userid,
+      mainPublicKey,
+      didType,
+      didValue,
+    });
+    const { signature } = await sendSignByWalletConnect(signContent, didValue.toLowerCase());
+    return {
+      signature,
+    };
+  };
+
+  const registerByWalletConnect = async (nickname?: string): Promise<void> => {
+    if (!userAccount) {
+      return;
+    }
+    const { address, userid, walletType } = userAccount;
+    const { publicKey, secretKey } = await getMainKeypairByWalletConnect({
+      password: confirmPassword.current,
+      did_value: address,
+      did_type: walletType,
+    });
+    const { signature } = await getRegisterSignContentByWalletConnect({
+      userid,
+      mainPublicKey: publicKey,
+      didType: walletType,
+      didValue: address,
+    });
+    await commonRegister({
+      mainPublicKey: publicKey,
+      mainPrivateKey: secretKey,
+      userid,
+      didType: walletType,
+      didValue: address,
+      signature: signature as string,
+      nickname,
+    });
+  };
+
+  const loginByWalletConnect = async () => {
+    if (!userAccount) {
+      return;
+    }
+    const { address, userid, walletType } = userAccount;
+    let localMainPrivateKey = '';
+    let localMainPublicKey = '';
+    if (mainKeys && address.toLowerCase() === mainKeys.walletAddress.toLowerCase()) {
+      localMainPrivateKey = mainKeys.privateKey;
+      localMainPublicKey = mainKeys.publicKey;
+    }
+    if (!localMainPublicKey || !localMainPrivateKey) {
+      const { publicKey, secretKey } = await getMainKeypairByWalletConnect({
+        password: confirmPassword.current,
+        did_value: address,
+        did_type: walletType,
+      });
+      localMainPrivateKey = secretKey;
+      localMainPublicKey = publicKey;
+    }
+
+    await commonLogin({
+      mainPrivateKey: localMainPrivateKey,
+      mainPublicKey: localMainPublicKey,
+      userid,
+      didType: walletType,
+      didValue: address,
+    });
+  };
+
   const getUserAccount = async (
     didType: WalletType = 'eth',
     address?: string,
@@ -207,19 +442,7 @@ const useLogin = (
       password: '',
       needJump: appType !== AppTypeEnum.pc,
     });
-    // await client.dappConnectClient.sendSign({
-    //   signContent,
-    //   didValue: address,
-    //   signType: signAuditType,
-    // });
   };
-
-  // useEffect(() => {
-  //   if (!mainKeys) {
-  //     return;
-  //   }
-  //   afterSignAndLogin();
-  // }, [mainKeys, registerSignRes]);
 
   const loginByQrCode = async () => {
     if (!userAccount) {
@@ -229,7 +452,6 @@ const useLogin = (
       if (!confirmPassword.current) {
         return;
       }
-      // setStep(StepStringEnum.LOGIN_SIGN_LOADING);
       await sendGetMainKeysSign(SignAuditTypeEnum.GET_KEYS_FOR_LOGIN);
       return;
     } else {
@@ -379,6 +601,12 @@ const useLogin = (
     web3MqSignCallback,
     setUserAccount,
     confirmPassword,
+    create,
+    connect,
+    closeModal,
+    onSessionConnected,
+    loginByWalletConnect,
+    registerByWalletConnect,
   };
 };
