+++ src/context/index.ts
@@ -4,6 +4,3 @@ export * from './ChatContext';
 export * from './ComponentContext';
 export * from './MessageContext';
 export * from './MessageInputContext';
-export * from './LoginContext';
-export * from './WalletConnectContext';
-export * from './BindDidConnectContext';
